size_t is not an int
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@18808 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/3D/FontManager.cxx
@@ -198,7 +198,7 @@ FTFont* BZFontFace_impl::loadSize(size_t size)
 
   setSize(size, font);
 
-  font->FaceSize(size);
+  font->FaceSize((unsigned int)size);
   bool doDisplayLists( ! BZDB.isTrue("NoDisplayListsForFonts"));
   font->UseDisplayList(doDisplayLists);
